requirements.txt
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.18
certifi==2015.4.28
Django==1.8.2
django-angular==0.7.15
django-appconf==1.0.1
django-cacheback==0.9.1
django-celery==3.1.16
django-compressor==1.5
django-debug-toolbar==1.3.2
django-filter==0.10.0
django-guardian==1.3
django-mptt==0.7.4
django-rest-swagger==0.3.2
djangorestframework==3.1.3
flower==0.8.2
futures==3.0.3
gevent==1.0.2
greenlet==0.4.7
kombu==3.0.26
pbr==1.2.0
Pillow==2.8.2
python-memcached==1.54
pytz==2015.4
PyYAML==3.11
requests==2.7.0
six==1.9.0
sqlparse==0.1.15
stevedore==1.5.0
tornado==4.2
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.6.0
wheel==0.24.0
